// bootstrap/compiler/src/generated/graph-capabilities.contracts.ts
export type GraphDeriveAdapter = {
	derive: (input: ModelTypes.GraphDeriveInput) => Promise<ModelTypes.DynamicDerivationResult>;
};